+++ requirements.txt
@@ -1,5 +1,6 @@
+PySide2==5.13.0
 ghostscript==0.6
 PyPDF2==1.26.0
-PyQt5==5.12.2
+PyQt5==5.13.0
 python-magic==0.4.15
 Wand==0.5.4
